Move error type
