+++ crates/uv-distribution/src/error.rs
@@ -8,11 +8,12 @@ use zip::result::ZipError;
 use crate::metadata::MetadataError;
 use uv_client::WrappedReqwestError;
 use uv_distribution_filename::WheelFilenameError;
-use uv_distribution_types::{AnyErrorBuild, IsBuildBackendError};
+use uv_distribution_types::IsBuildBackendError;
 use uv_fs::Simplified;
 use uv_normalize::PackageName;
 use uv_pep440::{Version, VersionSpecifiers};
 use uv_pypi_types::{HashAlgorithm, HashDigest, ParsedUrlError};
+use uv_types::AnyErrorBuild;
 
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
