+++ crates/uv-types/src/traits.rs
@@ -1,4 +1,6 @@
+use std::fmt::{Debug, Display, Formatter};
 use std::future::Future;
+use std::ops::Deref;
 use std::path::{Path, PathBuf};
 use uv_distribution_filename::DistFilename;
 
@@ -9,8 +11,8 @@ use uv_configuration::{
     BuildKind, BuildOptions, BuildOutput, ConfigSettings, LowerBound, SourceStrategy,
 };
 use uv_distribution_types::{
-    AnyErrorBuild, CachedDist, DependencyMetadata, IndexCapabilities, IndexLocations,
-    InstalledDist, IsBuildBackendError, Resolution, SourceDist,
+    CachedDist, DependencyMetadata, IndexCapabilities, IndexLocations, InstalledDist,
+    IsBuildBackendError, Resolution, SourceDist,
 };
 use uv_git::GitResolver;
 use uv_pep508::PackageName;
@@ -182,5 +184,63 @@ impl InstalledPackagesProvider for EmptyInstalledPackages {
 
     fn iter(&self) -> impl Iterator<Item = &InstalledDist> {
         std::iter::empty()
+    }
+}
+
+/// `anyhow::Error`-like wrapper type for [`BuildDispatch`] method return values, that also makes
+/// `IsBuildBackendError` work as `thiserror` `#[source]`.
+///
+/// The errors types have the same problem as [`BuildDispatch`] generally: The `uv-resolver`,
+/// `uv-installer` and `uv-build-frontend` error types all reference each other:
+/// Resolution and installation may need to build packages, while the build frontend needs to
+/// resolve and install for the PEP 517 build environment.
+///
+/// Usually, `anyhow::Error` is opaque error type of choice. In this case though, we error type
+/// that we can inspect on whether it's a build backend error with [`IsBuildBackendError`], and
+/// `anyhow::Error` does not allow attaching more traits. The next choice would be
+/// `Box<dyn std::error::Error + IsBuildFrontendError + Send + Sync + 'static>`, but `thiserror`
+/// complains about the internal `AsDynError` not being implemented when being used as `#[source]`.
+/// This struct is an otherwise transparent error wrapper that thiserror recognizes.
+pub struct AnyErrorBuild(Box<dyn IsBuildBackendError>);
+
+impl Debug for AnyErrorBuild {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        Debug::fmt(&self.0, f)
+    }
+}
+
+impl Display for AnyErrorBuild {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        Display::fmt(&self.0, f)
+    }
+}
+
+impl std::error::Error for AnyErrorBuild {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        self.0.source()
+    }
+
+    #[allow(deprecated)]
+    fn description(&self) -> &str {
+        self.0.description()
+    }
+
+    #[allow(deprecated)]
+    fn cause(&self) -> Option<&dyn std::error::Error> {
+        self.0.cause()
+    }
+}
+
+impl<T: IsBuildBackendError> From<T> for AnyErrorBuild {
+    fn from(err: T) -> Self {
+        Self(Box::new(err))
+    }
+}
+
+impl Deref for AnyErrorBuild {
+    type Target = dyn IsBuildBackendError;
+
+    fn deref(&self) -> &Self::Target {
+        &*self.0
     }
 }
